platform/tests: fix bug in nmtstp_ip_address_check_lifetime() calculating timeouts with unsigned integers
As one of the arguments in unsigned, the calculation is performed as
unsigned integers. That can actually lead to the wrong result. Fix it by
casting to the right (signed) types.

+++ src/platform/tests/test-common.c
@@ -955,8 +955,8 @@ nmtstp_ip_address_check_lifetime(const NMPlatformIPAddress *addr,
             if (adr != NM_PLATFORM_LIFETIME_PERMANENT)
                 return FALSE;
         } else {
-            if (adr - offset <= lft - CHECK_LIFETIME_MAX_DIFF
-                || adr - offset >= lft + CHECK_LIFETIME_MAX_DIFF)
+            if (((gint64) adr) - offset <= ((gint64) lft) - CHECK_LIFETIME_MAX_DIFF
+                || ((gint64) adr) - offset >= ((gint64) lft) + CHECK_LIFETIME_MAX_DIFF)
                 return FALSE;
         }
     }
